WIP: adjust 'first' test to true pattern
It looks like the primary artiact doesnt include the one-off baclup that
we used to begin iterating on this story .. this seems wrong. We need to
recreate this tarball to add it, then adopt the test to reference it
properly.

The second test modified here is failing, but correctly references the
backup. This should be used as an example, once the first test's
artifact is added to the larger tar file

Authored-by: Trevor Yacovone <tyacovone@pivotal.io>